backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-cloudinary-storage==0.3.0
cloudinary==1.36.0
psycopg2-binary==2.9.7
qrcode[pil]==7.4.2
Pillow==10.0.1
python-decouple==3.8
opencv-python==4.8.1.78
scikit-image==0.21.0
gunicorn==21.2.0
whitenoise==6.6.0
